[v4.1.0] px4ctrl_control: reload parameters & functions via `namespace`
